pyproject.toml
[tool.poetry]
name = "supernatural-soundtrack-scraper"
version = "0.1.0"
description = "Scrapes song data from the Supernatural Fandom wiki via the MediaWiki API and exports CSV and Excel."
authors = []
readme = "README.md"
packages = [{ include = "supernatural_soundtrack_scraper" }]

[tool.poetry.scripts]
scrape = "supernatural_soundtrack_scraper.__main__:main"

[tool.poetry.dependencies]
python = "^3.9"
requests = "*"
pandas = "*"
openpyxl = "*"
pydantic = "*"
pyyaml = "*"

[tool.poetry.group.dev.dependencies]
pytest = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
